Split measurement.cpp; add acpi_power_meter snapshot test
measurement.cpp now contains only the power_meter base class methods
and the power_meters vector (no heavy includes).

measurement_manager.cpp holds all manager/detection logic with the
full subtype include chain (acpi.h, extech.h, sysfs.h, opal-sensors.h,
parameters.h), following the same pattern as device_manager.cpp.

New test: tests/devices/test_acpi_power_meter.cpp covers all three
branches of acpi_power_meter::measure():
  - status != Discharging  → rate stays 0
  - power_now present      → rate = power_now / 1e6 (15.0 W)
  - fallback               → rate = current_now × voltage_now / 1e12 (24.0 W)

All 8 tests passing.
