+++ tests/devices/data/acpi_charging.ptrecord
@@ -1,0 +1,1 @@
+R /sys/class/power_supply/BAT0/status Q2hhcmdpbmcK
