+++ tests/devices/data/acpi_discharging_power_now.ptrecord
@@ -1,0 +1,2 @@
+R /sys/class/power_supply/BAT0/status RGlzY2hhcmdpbmcK
+R /sys/class/power_supply/BAT0/power_now MTUwMDAwMDAK
